+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Footer component can now contain an optional theme selector which requires the component to have JavaScript initialised
 - The Cookie library is now a singleton
+- Hero components have a new split layout option
 
 ### Changed